refactor: extract android creation helpers
