+++ src/managers/android/mod.rs
@@ -277,6 +277,7 @@
 //! This ensures consistent, predictable device ordering without hardcoded device lists.
 //!
 
+mod create;
 mod details;
 mod parser;
 mod sdk;
@@ -289,9 +290,7 @@ use crate::{
         files,
         keywords::{LOG_LEVEL_ERROR, LOG_LEVEL_FAILED},
         limits::{
-            ANDROID_COMMAND_PARTS_MINIMUM, MAX_DEVICE_NAME_CREATE_LENGTH,
-            MAX_DEVICE_NAME_PARTS_PROCESS, MAX_ERROR_MESSAGE_LENGTH, MIN_STRING_LENGTH_FOR_MATCH,
-            STORAGE_MB_TO_GB_DIVISOR, SYSTEM_IMAGE_PARTS_REQUIRED,
+            ANDROID_COMMAND_PARTS_MINIMUM, STORAGE_MB_TO_GB_DIVISOR, SYSTEM_IMAGE_PARTS_REQUIRED,
         },
         progress::{
             COMPLETION_THRESHOLD_PERCENTAGE, DOWNLOAD_PHASE_INCREMENT,
@@ -917,60 +916,6 @@ impl AndroidManager {
         }
     }
 
-    /// Get appropriate skin name for device type using dynamic lookup
-    async fn get_appropriate_skin(&self, device_id: &str, device_display: &str) -> Option<String> {
-        if device_id.is_empty() {
-            return None;
-        }
-
-        // First, try using the device ID directly as the skin name (most common case)
-        let primary_skin = device_id.to_string();
-
-        // Dynamically get available skins from SDK
-        let available_skins = self
-            .get_available_skins_from_sdk(device_id)
-            .await
-            .unwrap_or_default();
-
-        // 1. If there's a skin that exactly matches the device ID
-        if available_skins.iter().any(|skin| skin == &primary_skin) {
-            return Some(primary_skin);
-        }
-
-        // 2. Generate candidates from device display name and check
-        let display_based_skin = device_display
-            .split('(')
-            .next() // Only the part before parentheses
-            .unwrap_or(device_display)
-            .trim()
-            .replace(' ', "_")
-            .to_lowercase();
-
-        if available_skins
-            .iter()
-            .any(|skin| skin == &display_based_skin)
-        {
-            return Some(display_based_skin);
-        }
-
-        // 3. Search for skins with partial match
-        let device_lower = device_id.to_lowercase();
-        for skin in &available_skins {
-            let skin_lower = skin.to_lowercase();
-            // Main part of device ID is contained in skin name, or vice versa
-            if (device_lower.len() > MIN_STRING_LENGTH_FOR_MATCH
-                && skin_lower.contains(&device_lower))
-                || (skin_lower.len() > MIN_STRING_LENGTH_FOR_MATCH
-                    && device_lower.contains(&skin_lower))
-            {
-                return Some(skin.clone());
-            }
-        }
-
-        // 4. If all fail, return device ID as-is (handled by fallback strategy)
-        Some(primary_skin)
-    }
-
     /// Dynamically get available skins for device from SDK
     async fn get_available_skins_from_sdk(&self, _device_id: &str) -> Result<Vec<String>> {
         let mut skins = Vec::new();
@@ -1166,178 +1111,6 @@ impl DynamicDeviceProvider for AndroidManager {
 }
 
 impl AndroidManager {
-    /// Diagnose AVD creation issues and provide specific solutions
-    pub async fn diagnose_avd_creation_issues(&self, config: &DeviceConfig) -> Result<String> {
-        let mut diagnosis = Vec::new();
-
-        // Check 1: Android SDK availability
-        diagnosis.push("=== Android SDK Diagnosis ===".to_string());
-
-        // Check 2: Available system images
-        let available_images = self.list_available_system_images().await?;
-        diagnosis.push(format!(
-            "Available system images: {}",
-            available_images.len()
-        ));
-        if available_images.is_empty() {
-            diagnosis.push("❌ No system images found! Install with: sdkmanager \"system-images;android-XX;google_apis_playstore;arm64-v8a\"".to_string());
-        } else {
-            diagnosis.push("✅ System images available".to_string());
-            diagnosis.push(format!(
-                "First 3: {}",
-                available_images
-                    .iter()
-                    .take(MAX_DEVICE_NAME_PARTS_PROCESS)
-                    .cloned()
-                    .collect::<Vec<_>>()
-                    .join(", ")
-            ));
-        }
-
-        // Check 3: Available device types
-        let available_devices = self.list_available_devices().await?;
-        diagnosis.push(format!(
-            "Available device types: {}",
-            available_devices.len()
-        ));
-        if available_devices.is_empty() {
-            diagnosis.push("❌ No device types found! Check Android SDK installation".to_string());
-        } else {
-            diagnosis.push("✅ Device types available".to_string());
-            diagnosis.push(format!(
-                "First 3: {}",
-                available_devices
-                    .iter()
-                    .take(MAX_DEVICE_NAME_PARTS_PROCESS)
-                    .map(|(id, display)| format!("{display} ({id})"))
-                    .collect::<Vec<_>>()
-                    .join(", ")
-            ));
-        }
-
-        // Check 4: Specific system image for this config
-        let (tag, abi) = if let Some((found_tag, found_abi)) = self
-            .get_first_available_system_image(&config.version)
-            .await?
-        {
-            (found_tag, found_abi)
-        } else {
-            (
-                "google_apis_playstore".to_string(),
-                defaults::default_abi().to_string(),
-            )
-        };
-
-        let package_path = format!(
-            "system-images;android-{version};{tag};{abi}",
-            version = config.version,
-            tag = tag,
-            abi = abi
-        );
-        let image_available = self
-            .check_system_image_available(&config.version, &tag, &abi)
-            .await
-            .unwrap_or(false);
-
-        diagnosis.push(format!("Required system image: {package_path}"));
-        if image_available {
-            diagnosis.push("✅ Required system image is available".to_string());
-        } else {
-            diagnosis.push("❌ Required system image NOT available".to_string());
-            diagnosis.push(format!("Install with: sdkmanager \"{package_path}\""));
-        }
-
-        // Check 5: Device type availability
-        let device_id = Self::find_matching_device_id(&available_devices, &config.device_type);
-        diagnosis.push(format!(
-            "Required device type: {} ({})",
-            config.device_type,
-            device_id.as_deref().unwrap_or("NOT FOUND")
-        ));
-        if device_id.is_some() {
-            diagnosis.push("✅ Required device type is available".to_string());
-        } else {
-            diagnosis.push("❌ Required device type NOT found".to_string());
-            diagnosis.push("Suggestion: Use one of the available device types above".to_string());
-        }
-
-        Ok(diagnosis.join("\n"))
-    }
-
-    /// Find matching device ID from available devices list
-    fn find_matching_device_id(
-        available_devices: &[(String, String)],
-        device_type: &str,
-    ) -> Option<String> {
-        // Try exact ID match first
-        if let Some((id, _)) = available_devices.iter().find(|(id, _)| id == device_type) {
-            return Some(id.clone());
-        }
-
-        // Try exact display name match
-        if let Some((id, _)) = available_devices
-            .iter()
-            .find(|(_, display)| display == device_type)
-        {
-            return Some(id.clone());
-        }
-
-        // Try partial match for display names (handles variations like quotes)
-        let cleaned_config = device_type
-            .chars()
-            .filter(|c| c.is_alphanumeric() || c.is_whitespace())
-            .collect::<String>()
-            .to_lowercase();
-
-        // More flexible matching
-        available_devices.iter().find_map(|(id, display)| {
-            let cleaned_display = display
-                .chars()
-                .filter(|c| c.is_alphanumeric() || c.is_whitespace())
-                .collect::<String>()
-                .to_lowercase();
-
-            // Exact match
-            if cleaned_config == cleaned_display {
-                return Some(id.clone());
-            }
-
-            // Match by main keywords contained in device type name
-            let config_words: Vec<&str> = cleaned_config.split_whitespace().collect();
-            let display_words: Vec<&str> = cleaned_display.split_whitespace().collect();
-
-            // Check if main keywords match (e.g., "Galaxy", "Pixel", "Nexus")
-            let important_words = ["galaxy", "pixel", "nexus", "tv", "wear", "automotive"];
-            for word in &important_words {
-                if cleaned_config.contains(word) && cleaned_display.contains(word) {
-                    // Further detailed check (e.g., "Galaxy S24" vs "Galaxy Nexus")
-                    let config_specific: Vec<&str> = config_words
-                        .iter()
-                        .filter(|w| w.chars().any(|c| c.is_ascii_digit()) || w.len() > 4)
-                        .cloned()
-                        .collect();
-                    let display_specific: Vec<&str> = display_words
-                        .iter()
-                        .filter(|w| w.chars().any(|c| c.is_ascii_digit()) || w.len() > 4)
-                        .cloned()
-                        .collect();
-
-                    if !config_specific.is_empty() && !display_specific.is_empty() {
-                        // Only when specific keywords match
-                        if config_specific.iter().any(|w| display_specific.contains(w)) {
-                            return Some(id.clone());
-                        }
-                    } else if config_specific.is_empty() && display_specific.is_empty() {
-                        // Basic matching when both have generic names
-                        return Some(id.clone());
-                    }
-                }
-            }
-
-            None
-        })
-    }
-
     async fn get_available_tags_for_api_level(&self, api_level: u32) -> Result<Vec<String>> {
         let images = self.list_available_system_images().await?;
         let mut tags = HashSet::new();
@@ -1574,237 +1347,7 @@ impl DeviceManager for AndroidManager {
     }
 
     async fn create_device(&self, config: &DeviceConfig) -> Result<()> {
-        // AVD names must follow strict rules: only a-z A-Z 0-9 . _ - are allowed
-        // But preserve the original name for display purposes in config.ini
-        let safe_name = config
-            .name
-            .chars()
-            .filter_map(|c| match c {
-                // Keep only allowed characters, replace spaces and others with underscores
-                c if c.is_ascii_alphanumeric() || c == '.' || c == '-' => Some(c),
-                ' ' | '_' => Some('_'), // Convert spaces to underscores
-                _ => None,              // Remove all other characters
-            })
-            .collect::<String>()
-            .trim_matches('_') // Remove leading/trailing underscores
-            .to_string();
-
-        if safe_name.is_empty() {
-            return Err(anyhow::anyhow!(
-                "Device name '{}' contains only invalid characters and cannot be used for AVD creation.",
-                config.name
-            ));
-        }
-
-        // Check if device with this name already exists
-        let existing_devices = self.list_devices().await?;
-        if existing_devices.iter().any(|d| d.name == safe_name) {
-            return Err(anyhow::anyhow!(
-                "Device with name '{safe_name}' already exists. Please choose a different name or delete the existing device first."
-            ));
-        }
-
-        // Try to find available system image for this API level
-        let (tag, abi) = if let Some((found_tag, found_abi)) = self
-            .get_first_available_system_image(&config.version)
-            .await?
-        {
-            (found_tag, found_abi)
-        } else {
-            // If no system image found, try default values
-            let default_tag = config
-                .additional_options
-                .get("tag")
-                .map_or("google_apis_playstore", |s| s.as_str());
-            let default_abi = config
-                .additional_options
-                .get("abi")
-                .map_or(defaults::default_abi(), |s| s.as_str());
-            (default_tag.to_string(), default_abi.to_string())
-        };
-
-        let package_path = format!(
-            "system-images;android-{};{};{}",
-            config.version, // API level
-            tag,
-            abi
-        );
-
-        // Check if system image is available
-        let image_available = self
-            .check_system_image_available(&config.version, &tag, &abi)
-            .await
-            .unwrap_or(false);
-
-        if !image_available {
-            let available_images = self.list_available_system_images().await?;
-            return Err(anyhow::anyhow!(
-                "System image '{}' not found. Install it with: sdkmanager \"{}\"\nAvailable images: {}",
-                package_path, package_path, available_images.join(", ")
-            ));
-        }
-
-        // **IMPROVED APPROACH**: Use proper device and skin parameters
-        let mut args = vec!["create", "avd", "-n", &safe_name, "-k", &package_path];
-
-        // Add device parameter if valid - use ID for better compatibility
-        let device_param =
-            if !config.device_type.is_empty() && config.device_type.to_lowercase() != "custom" {
-                let available_devices = self.list_available_devices().await?;
-                Self::find_matching_device_id(&available_devices, &config.device_type)
-            } else {
-                None
-            };
-
-        // Use default if device parameter is not found
-        if let Some(ref device_id) = device_param {
-            args.push("--device");
-            args.push(device_id);
-        } else {
-            // Omit device parameter - avdmanager will use default device
-            log::warn!(
-                "Device type '{}' not found, using default device",
-                config.device_type
-            );
-        }
-
-        // Dynamically get and specify skin (use fallback strategy on error)
-        let skin_name = if let Some(ref device_id) = device_param {
-            self.get_appropriate_skin(device_id, &config.device_type)
-                .await
-        } else {
-            self.get_appropriate_skin(&config.device_type, &config.device_type)
-                .await
-        };
-
-        if let Some(ref skin) = skin_name {
-            args.push("--skin");
-            args.push(skin);
-        }
-
-        let result = self
-            .command_executor
-            .run(&self.avdmanager_path, &args)
-            .await;
-
-        // Retry without skin if skin error occurs
-        let result = if result.is_err() && skin_name.is_some() {
-            let error_str = result.as_ref().unwrap_err().to_string();
-            if error_str.to_lowercase().contains("skin") {
-                log::warn!(
-                    "Skin '{}' failed, retrying without skin",
-                    skin_name.as_ref().unwrap()
-                );
-                // Retry after removing skin parameter
-                let mut fallback_args =
-                    vec!["create", "avd", "-n", &safe_name, "-k", &package_path];
-                if let Some(ref device_id) = device_param {
-                    fallback_args.push("--device");
-                    fallback_args.push(device_id);
-                }
-                self.command_executor
-                    .run(&self.avdmanager_path, &fallback_args)
-                    .await
-            } else {
-                result
-            }
-        } else {
-            result
-        };
-
-        match result {
-            Ok(_output) => {
-                // **OPTIMIZED**: Only update config if absolutely necessary
-                // The avdmanager should have created a good base configuration with --device and --skin
-                // We only need to fine-tune specific settings
-                if let Err(e) = self
-                    .fine_tune_avd_config(&safe_name, config, &tag, &abi)
-                    .await
-                {
-                    // Log warning but don't fail the entire operation
-                    eprintln!("Warning: Failed to fine-tune AVD configuration: {e}");
-                }
-
-                Ok(())
-            }
-            Err(e) => {
-                // Provide comprehensive error diagnostics
-                let error_str = e.to_string();
-
-                // Create ultra-compact diagnostic info for UI
-                let mut diagnostic_info = Vec::new();
-                // Shortened name (max 20 characters)
-                let short_name = if safe_name.len() > MAX_DEVICE_NAME_CREATE_LENGTH {
-                    format!(
-                        "{name}...",
-                        name = &safe_name[..MAX_DEVICE_NAME_CREATE_LENGTH - 3]
-                    )
-                } else {
-                    safe_name.clone()
-                };
-                diagnostic_info.push(format!("AVD: {short_name}"));
-                diagnostic_info.push(format!("API: {}", config.version));
-
-                // Create concise error message with important information at the beginning
-                if error_str.contains("system image")
-                    || error_str.contains("package path")
-                    || error_str.contains("not installed")
-                {
-                    Err(anyhow::anyhow!(
-                        "System image not installed for API {}\nRun: sdkmanager \"{}\"",
-                        config.version,
-                        package_path
-                    ))
-                } else if error_str.contains("license") || error_str.contains("accept") {
-                    Err(anyhow::anyhow!(
-                        "Android SDK licenses not accepted\nRun: sdkmanager --licenses"
-                    ))
-                } else if error_str.contains("already exists") {
-                    Err(anyhow::anyhow!(
-                        "AVD '{}' already exists\nDelete existing or choose different name",
-                        config.name
-                    ))
-                } else if error_str.contains("device") && error_str.contains("not found") {
-                    Err(anyhow::anyhow!(
-                        "Device type '{}' not found\nCheck available types in device list",
-                        config.device_type
-                    ))
-                } else {
-                    // Generic error - only the most important information
-                    let key_error = if error_str.contains("Error:") {
-                        error_str
-                            .split("Error:")
-                            .nth(1)
-                            .unwrap_or(&error_str)
-                            .trim()
-                    } else if error_str.contains("failed") {
-                        error_str
-                            .split("failed")
-                            .nth(0)
-                            .unwrap_or(&error_str)
-                            .trim()
-                    } else {
-                        &error_str
-                    };
-
-                    let short_error = if key_error.len() > MAX_ERROR_MESSAGE_LENGTH {
-                        format!(
-                            "{error}...",
-                            error = &key_error[..MAX_ERROR_MESSAGE_LENGTH - 3]
-                        )
-                    } else {
-                        key_error.to_string()
-                    };
-
-                    Err(anyhow::anyhow!(
-                        "AVD creation failed: {}\nAVD: {} | API: {}",
-                        short_error,
-                        short_name,
-                        config.version
-                    ))
-                }
-            }
-        }
+        self.create_device_internal(config).await
     }
 
     async fn delete_device(&self, identifier: &str) -> Result<()> {
